requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
PyJWT
python-dotenv
sqlparse
typing_extensions
drf-yasg
django-redis==3.1.6
django-simple-history
minio
django-storages==1.14.4
boto3==1.35.16
botocore==1.35.16
django-browser-reload~=1.13.0
python-decouple~=3.8
gunicorn>=21.2.0
setuptools
pillow
cryptography
requests
